+++ src/commands/spconfig-export/SPConfigExporter.ts
@@ -74,11 +74,17 @@ export class SPConfigExporter {
             fs.writeFileSync(this.target, JSON.stringify(data, null, 2), { encoding: "utf8" });
         } else {
             for (let obj of data.objects) {
+                let name = obj.self.name
+                // names for lifecycle state are not unique. Need to prefix with identity profile name
+                if ("LIFECYCLE_STATE" === obj.self.type) {
+                    name = `${obj.object.identityProfileRef.name}_${name}`
+                }
+
                 const targetFilename = PathProposer.getSPConfigMultipeFileFilename(
                     this.tenantName as string,
                     this.tenantDisplayName as string,
                     obj.self.type,
-                    sanitizeFilename(obj.self.name,)
+                    sanitizeFilename(name)
                 );
 
                 const targetFilepath = path.join(this.target, targetFilename);